pyproject.toml
# sauron: bot for monitoring telos block producer
# Copyright (C) 2024-present (in stewardship of openrepublic)
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/settings/#lint_ignore
ignore = []

[project]
name = "sauron"
version = "0.1.0"
description = "bot for monitoring telos block producer"
authors = [{ name = "Nelson Torres", email = "nelson.torres.a@proton.me" }]
requires-python = ">=3.12, <3.13"
license = "AGPL-3.0-or-later"
readme = "README.md"
keywords = [
    "async",
]
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
]

dependencies = [
    'aiohttp==3.9.5, <4.0.0',
    'anyio >=3.7.1, <4.0.0',
    'click >=7.1.2, <8.0.0',
    'msgspec >=0.18.6, <0.19.0',
    'ntplib >=0.4.0, <1.0.0',
    'pyTelegramBotAPI >=4.21.0, <5.0.0',
    'requests >=2.31.0, <3.0.0',
    'speedtest-cli >=2.1.3, <3.0.0',
    'telebot >=0.0.5, <1.0.0',
    'urllib3 >=2.2.2, <3.0.0',
    'py-leap',
]


[dependency-groups]
dev = [
    "pytest >=6.0.0, <7.0.0",
]

[project.scripts]
sauron = "sauron.cli:sauron"

[tool.hatch.build.targets.sdist]
include = ["sauron"]

[tool.hatch.build.targets.wheel]
include = ["sauron"]

[tool.uv.sources]
py-leap = { git = "https://github.com/guilledk/py-leap.git", tag = "v0.1a25" }
